requirements.txt
crewai>=0.80.0
crewai-tools>=0.17.0
pandas>=2.0.0
ta>=0.11.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
vaderSentiment>=3.3.2
pycoingecko>=3.1.0
python-binance>=1.0.19
colorama>=0.4.6
rich>=13.0.0
yfinance>=0.2.40
